Addition of Insertion Sort Algorithm
Insertion Sort is an algorithm which takes a list as an input and
returns a sorted version of the list. The sorting is achieved by
assigning a jth value of the list to a variable named temp and compares
this to the other values within the list while A[i] > temp.